+++ ChangeLog.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Haskell Package Versioning policy](https://pvp.haskell.org/).
 
 ## Unreleased changes
+- Add 'Container'
 - Add 'Updatable' instance for Container
 - Add 'updateContainer' to update Container with different updater for each Shgif
 - 'Shgif.Type' will export Lenses for Container